backend/requirements-minimal.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
openai
psutil
requests
